# website/static/json/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "couic",
        "description": "A lightweight eBPF-powered network filtering solution controllable through a REST API",
        "license": {
            "name": "MIT OR Apache-2.0",
            "identifier": "MIT OR Apache-2.0"
        },
        "version": "0.5.12"
    },
...
    "paths": {
...
        "/v1/client": {
            "get": {
                "tags": [
                    "clients"
                ],
                "summary": "List all clients",
                "operationId": "list_clients",
                "responses": {
                    "200": {
                        "description": "List of clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "clients"
                ],
                "summary": "Create a new client",
                "operationId": "create_client",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Client created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (malformed JSON)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Client already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/v1/client/{name}": {
            "get": {
                "tags": [
                    "clients"
                ],
                "summary": "Get a client by name",
                "operationId": "get_client",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Client name (e.g., 'client1')",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "examples": [
                                "client1"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid path parameter)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "clients"
                ],
                "summary": "Delete a client by name",
                "operationId": "delete_client",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Client name (e.g., 'client1')",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "examples": [
                                "client1"
                            ]
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Client deleted"
                    },
                    "400": {
                        "description": "Bad request (invalid path parameter)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompositeError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CompositeError": {
                "type": "object",
                "required": [
                    "code",
                    "message",
                    "errors"
                ],
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        },
                        "propertyNames": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "ErrorCode": {
                "type": "string",
                "description": "Enum representing error codes used in the API",
                "enum": [
                    "Eprocessing",
                    "Eunauthorized",
                    "Enotfound",
                    "Econflict",
                    "Ebadrequest",
                    "Einvalid",
                    "Einternal",
                    "Enotimplemented"
                ]
            },
            "ErrorDetail": {
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "$ref": "#/components/schemas/ErrorCode"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "RbacGroup": {
                "type": "string",
                "description": "RBAC group defining client permissions",
                "enum": [
                    "admin",
                    "clientro",
                    "clientrw",
                    "monitoring",
                    "peering"
                ]
            },
...
            "Client": {
                "type": "object",
                "description": "Client information (returned when listing clients)",
                "required": [
                    "name",
                    "group"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Client name",
                        "maxLength": 48,
                        "pattern": "^[a-zA-Z0-9._-]+$",
                        "examples": [
                            "prometheus"
                        ]
                    },
                    "group": {
                        "$ref": "#/components/schemas/RbacGroup"
                    }
                }
            },
            "ClientResponse": {
                "type": "object",
                "description": "Client information with token (returned when creating a client)",
                "required": [
                    "name",
                    "group",
                    "token"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Client name",
                        "maxLength": 48,
                        "pattern": "^[a-zA-Z0-9._-]+$",
                        "examples": [
                            "prometheus"
                        ]
                    },
                    "group": {
                        "$ref": "#/components/schemas/RbacGroup"
                    },
                    "token": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Bearer token for authentication",
                        "examples": [
                            "d6ac883a-8050-4408-bf1e-5b07e9965191"
                        ]
                    }
                }
            },
            "ClientRequest": {
                "type": "object",
                "description": "Request body for creating a client",
                "required": [
                    "name",
                    "group"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Client name",
                        "maxLength": 48,
                        "pattern": "^[a-zA-Z0-9._-]+$",
                        "examples": [
                            "prometheus"
                        ]
                    },
                    "group": {
                        "$ref": "#/components/schemas/RbacGroup"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Uuidv4",
                "description": "An API key is a token that you provide when making API calls. Include the token in a header parameter called `Authorization`.\nExample: `Authorization: Bearer b03dd133-3827-4793-b0df-37f6ffbdd9a9`"
            }
        }
    },
...
}